restrict list access to roles with Pundit policy

+++ app/controllers/lists_controller.rb
@@ -1,30 +1,36 @@
 class ListsController < ApplicationController
   before_action :set_list, only: [:show, :edit, :update, :destroy]
+  after_action :verify_authorized
 
   # GET /lists
   # GET /lists.json
   def index
     @lists = List.includes(:user).all
+    authorize List
   end
 
   # GET /lists/1
   # GET /lists/1.json
   def show
+    authorize @list
   end
 
   # GET /lists/new
   def new
     @list = List.new
+    authorize @list
   end
 
   # GET /lists/1/edit
   def edit
+    authorize @list
   end
 
   # POST /lists
   # POST /lists.json
   def create
     @list = List.new(list_params)
+    authorize @list
     @list.user = current_user
 
     respond_to do |format|
@@ -41,6 +47,7 @@ class ListsController < ApplicationController
   # PATCH/PUT /lists/1
   # PATCH/PUT /lists/1.json
   def update
+    authorize @list
     respond_to do |format|
       if @list.update(list_params)
         format.html { redirect_to @list, notice: 'List was successfully updated' }
@@ -55,6 +62,7 @@ class ListsController < ApplicationController
   # DELETE /lists/1
   # DELETE /lists/1.json
   def destroy
+    authorize @list
     @list.destroy
     respond_to do |format|
       format.html { redirect_to lists_url, notice: 'List was successfully destroyed' }
@@ -64,6 +72,7 @@ class ListsController < ApplicationController
 
   def move_album
     list = List.find(params[:list_id])
+    authorize list
     album = Album.find(params[:album_id])
     list.move_album(album, params[:direction].to_sym)
     list.save!
@@ -72,6 +81,7 @@ class ListsController < ApplicationController
 
   def delete_album
     list = List.find(params[:list_id])
+    authorize list
     album = Album.find(params[:album_id])
     list.delete_album(album)
     list.save!
